feat: 지도 페이지 오픈그래프 적용을 위해 서버/클라이언트 컴포넌트 분리(#192)
